fix: mixins are not ClassElement
This PR solves the issue but i am not sure if this is correct approach.
I would like to add tests if someone could point me where would be the
best place.

I will mention @ykmnkmi as this is mostly his work.

When compiling this:

```
@Directive(selector: 'dir')
class MyDir with MyMixin {}

mixin MyMixin {}
```

you get this error:
```
type 'MixinElementImpl' is not a subtype of type 'ClassElement' in type cast.
```